+++ philo/setters.c
@@ -6,7 +6,7 @@
 /*   By: asayad <asayad@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/07 07:30:29 by asayad            #+#    #+#             */
-/*   Updated: 2024/07/29 16:00:26 by asayad           ###   ########.fr       */
+/*   Updated: 2024/08/01 09:44:35 by asayad           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	print_msg(t_philo *p_d, char *msg)
 {
-	u_int64_t	time;
+	long long	time;
 
 	pthread_mutex_lock(&p_d->data->mut_print);
 	time = get_current_time() - g_start_t(p_d);
@@ -28,10 +28,3 @@ void	set_iter(t_data *data, bool status)
 	data->iter = status;
 	pthread_mutex_unlock(&data->mut_iter);
 }
-
-// void	set_phil_state(t_philo *p_d, t_state state)
-// {
-// 	// pthread_mutex_lock(&p_d->data->mut_phil_state);
-// 	// p_d->phil_state = state;
-// 	// pthread_mutex_unlock(&p_d->data->mut_phil_state);
-// }
